se lee informacion por el websocktet

+++ Capa-Presentacion/EnviosApp/app/src/main/java/com/example/envios_app/utils/DistanceUtils.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 public class DistanceUtils {
     private final static double AVERAGE_RADIUS_OF_EARTH_KM = 6371;
 
-    public static int calculateDistanceInKilometer(double userLat, double userLng, double venueLat, double venueLng) {
+    public static double calculateDistanceInKilometer(double userLat, double userLng, double venueLat, double venueLng) {
 
         double latDistance = Math.toRadians(userLat - venueLat);
         double lngDistance = Math.toRadians(userLng - venueLng);
@@ -19,7 +19,7 @@ public class DistanceUtils {
 
         double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
 
-        return (int) Math.round(AVERAGE_RADIUS_OF_EARTH_KM * c);
+        return Math.round(AVERAGE_RADIUS_OF_EARTH_KM * c);
     }
 
     public static LatLng moveLatLngInKilometer(double latMove, double lngMove, LatLng position){
